fix(metrics): fix the bug where user deletion event won't update HubSpot
